requirements.txt
bluepy
python-logging-loki